backend/requirements.txt
accelerate==1.3.0
alembic==1.14.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
art==6.4
asyncpg==0.30.0
bcrypt==3.2.0
billiard==4.2.1
boto3==1.36.19
botocore==1.36.19
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crcmod==1.7
cryptography==44.0.0
datasketch==1.6.5
distro==1.9.0
dnspython==2.7.0
duckdb==1.2.0
ecdsa==0.19.0
et_xmlfile==2.0.0
fastapi==0.104.1
filelock==3.17.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jmespath==0.10.0
jsondiff==2.2.1
kombu==5.4.2
Levenshtein==0.26.1
Mako==1.3.9
MarkupSafe==3.0.2
motor==3.7.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.61.1
openpyxl==3.1.5
oss2==2.19.1
packaging==24.2
pandas==2.2.3
passlib==1.7.4
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.11
pymssql==2.3.2
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-Levenshtein==0.26.1
python-multipart==0.0.6
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
s3transfer==0.11.2
safetensors==0.5.2
scipy==1.15.1
sentencepiece==0.2.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
sseclient==0.0.27
sseclient-py==1.8.0
starlette==0.27.0
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.24.0
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
